yaml parser:
Added strict decoding of yaml to raise error on unknown fields.
trend functionality:

Added note that reversetrend can cause offset depending on function.
Added parameter of PeriodDuration,
this is used in place of T for trendAnomaly maths functions if set.
If it is not set (i.e. 0) it will use Duration.
general:

Added and/or updated relevant test cases